Extract flag styles to Flag component
